DynaMite Implementation with Entity and FSM
Did a second version of DynaMite.
The previous version was done before the implementation of State Machine and Entity system, so well, now it is done with both.
There are things I cannot do right now as I do not see a break meter anywhere and NavigationAgent is not working as navigationMap is not available I think? In any case, I wrote code on my own instead of using functions from entity.gd and enemy.gd where that's impossible.